584 redirect login (#633)
* Redirect Login to Home

* Fix user is login
