+++ frontend/src/components/atoms/Button.jsx
@@ -89,7 +89,7 @@ function Button({
 
 Button.propTypes = {
   type: PropTypes.string.isRequired,
-  text: PropTypes.string,
+  text: PropTypes.oneOfType([PropTypes.object, PropTypes.string]),
   loadingText: PropTypes.string,
   isLoading: PropTypes.bool,
   iconPosition: PropTypes.string,
